backend/requirements.txt
anyio==3.6.1
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.4
py==1.11.0
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
